Fixed betting so that it is now a double value instead of an int.

+++ game_server/src/blackjack/server/Blackjack.java
@@ -20,7 +20,7 @@ public class Blackjack {
 	private List<BlackjackPlayer> _toRemove = new ArrayList<BlackjackPlayer>();
 	private Dealer _dealer;
 	private Shoe _shoe;
-	private int bet;
+	private double bet;
 	
 	private static final int maxPlayers = 6;
